# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  code: string;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  admin: "admin",
  user: "user",
} as const;

export interface LoginResponse {
  token: string;
  role: LoginResponseRole;
  message: string;
}

export interface AccessCode {
  id: number;
  code: string;
  used: boolean;
  usedAt?: string | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type StreamStatus = (typeof StreamStatus)[keyof typeof StreamStatus];

export const StreamStatus = {
  idle: "idle",
  streaming: "streaming",
  error: "error",
} as const;

export type StreamActiveKey =
  (typeof StreamActiveKey)[keyof typeof StreamActiveKey];

export const StreamActiveKey = {
  primary: "primary",
  backup: "backup",
} as const;

export interface Stream {
  id: number;
  name: string;
  sourceUrl: string;
  primaryStreamKey: string;
  backupStreamKey?: string | null;
  rtmpsUrl: string;
  status: StreamStatus;
  activeKey: StreamActiveKey;
  switchInterval: number;
  copyrightProtection: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStreamRequest {
  name: string;
  sourceUrl: string;
  primaryStreamKey: string;
  backupStreamKey?: string;
  rtmpsUrl: string;
  switchInterval?: number;
  copyrightProtection?: boolean;
}

export interface UpdateStreamRequest {
  name?: string;
  sourceUrl?: string;
  primaryStreamKey?: string;
  backupStreamKey?: string;
  rtmpsUrl?: string;
  switchInterval?: number;
  copyrightProtection?: boolean;
}

export type StreamStatusResponseStatus =
  (typeof StreamStatusResponseStatus)[keyof typeof StreamStatusResponseStatus];

export const StreamStatusResponseStatus = {
  idle: "idle",
  streaming: "streaming",
  error: "error",
} as const;

export type StreamStatusResponseActiveKey =
  (typeof StreamStatusResponseActiveKey)[keyof typeof StreamStatusResponseActiveKey];

export const StreamStatusResponseActiveKey = {
  primary: "primary",
  backup: "backup",
} as const;

export interface StreamStatusResponse {
  id: number;
  status: StreamStatusResponseStatus;
  activeKey: StreamStatusResponseActiveKey;
  message: string;
  pid?: number | null;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}