# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Facebook Live Streamer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: admin
    description: Admin operations
  - name: streams
    description: Stream management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify:
    get:
      operationId: verifySession
      tags: [auth]
      summary: Verify current session
      responses:
        "200":
          description: Session valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/codes:
    get:
      operationId: getAccessCodes
      tags: [admin]
      summary: List all access codes
      responses:
        "200":
          description: List of access codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: generateAccessCode
      tags: [admin]
      summary: Generate a new access code
      responses:
        "201":
          description: Code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
  /admin/codes/{id}:
    delete:
      operationId: deleteAccessCode
      tags: [admin]
      summary: Delete an access code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Code deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /streams:
    get:
      operationId: getStreams
      tags: [streams]
      summary: Get all streams
      responses:
        "200":
          description: List of streams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
    post:
      operationId: createStream
      tags: [streams]
      summary: Create a new stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStreamRequest"
      responses:
        "201":
          description: Stream created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
  /streams/{id}:
    get:
      operationId: getStream
      tags: [streams]
      summary: Get stream by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
        "404":
          description: Not found
    put:
      operationId: updateStream
      tags: [streams]
      summary: Update stream
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStreamRequest"
      responses:
        "200":
          description: Stream updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
    delete:
      operationId: deleteStream
      tags: [streams]
      summary: Delete stream
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /streams/{id}/start:
    post:
      operationId: startStream
      tags: [streams]
      summary: Start streaming
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatusResponse"
  /streams/{id}/stop:
    post:
      operationId: stopStream
      tags: [streams]
      summary: Stop streaming
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatusResponse"
  /streams/{id}/status:
    get:
      operationId: getStreamStatus
      tags: [streams]
      summary: Get stream status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatusResponse"
  /streams/{id}/switch-key:
    post:
      operationId: switchStreamKey
      tags: [streams]
      summary: Switch between primary and backup stream keys
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Key switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatusResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    LoginRequest:
      type: object
      properties:
        code:
          type: string
      required: [code]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [admin, user]
        message:
          type: string
      required: [token, role, message]
    AccessCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        used:
          type: boolean
        usedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, code, used, createdAt]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    Stream:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sourceUrl:
          type: string
        primaryStreamKey:
          type: string
        backupStreamKey:
          type: string
          nullable: true
        rtmpsUrl:
          type: string
        status:
          type: string
          enum: [idle, streaming, error]
        activeKey:
          type: string
          enum: [primary, backup]
        switchInterval:
          type: integer
        copyrightProtection:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, sourceUrl, primaryStreamKey, rtmpsUrl, status, activeKey, switchInterval, copyrightProtection, createdAt, updatedAt]
    CreateStreamRequest:
      type: object
      properties:
        name:
          type: string
        sourceUrl:
          type: string
        primaryStreamKey:
          type: string
        backupStreamKey:
          type: string
        rtmpsUrl:
          type: string
          default: rtmps://live-api-s.facebook.com:443/rtmp/
        switchInterval:
          type: integer
          default: 0
        copyrightProtection:
          type: boolean
          default: true
      required: [name, sourceUrl, primaryStreamKey, rtmpsUrl]
    UpdateStreamRequest:
      type: object
      properties:
        name:
          type: string
        sourceUrl:
          type: string
        primaryStreamKey:
          type: string
        backupStreamKey:
          type: string
        rtmpsUrl:
          type: string
        switchInterval:
          type: integer
        copyrightProtection:
          type: boolean
    StreamStatusResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [idle, streaming, error]
        activeKey:
          type: string
          enum: [primary, backup]
        message:
          type: string
        pid:
          type: integer
          nullable: true
      required: [id, status, activeKey, message]
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]