# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with access code
 */
export const LoginBody = zod.object({
  code: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  role: zod.enum(["admin", "user"]),
  message: zod.string(),
});

/**
 * @summary Verify current session
 */
export const VerifySessionResponse = zod.object({
  token: zod.string(),
  role: zod.enum(["admin", "user"]),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary List all access codes
 */
export const GetAccessCodesResponseItem = zod.object({
  id: zod.number(),
  code: zod.string(),
  used: zod.boolean(),
  usedAt: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetAccessCodesResponse = zod.array(GetAccessCodesResponseItem);

/**
 * @summary Delete an access code
 */
export const DeleteAccessCodeParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAccessCodeResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get all streams
 */
export const GetStreamsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  sourceUrl: zod.string(),
  primaryStreamKey: zod.string(),
  backupStreamKey: zod.string().nullish(),
  rtmpsUrl: zod.string(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  switchInterval: zod.number(),
  copyrightProtection: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetStreamsResponse = zod.array(GetStreamsResponseItem);

/**
 * @summary Create a new stream
 */
export const createStreamBodyRtmpsUrlDefault = `rtmps://live-api-s.facebook.com:443/rtmp/`;
export const createStreamBodySwitchIntervalDefault = 0;
export const createStreamBodyCopyrightProtectionDefault = true;

export const CreateStreamBody = zod.object({
  name: zod.string(),
  sourceUrl: zod.string(),
  primaryStreamKey: zod.string(),
  backupStreamKey: zod.string().optional(),
  rtmpsUrl: zod.string().default(createStreamBodyRtmpsUrlDefault),
  switchInterval: zod.number().default(createStreamBodySwitchIntervalDefault),
  copyrightProtection: zod
    .boolean()
    .default(createStreamBodyCopyrightProtectionDefault),
});

/**
 * @summary Get stream by ID
 */
export const GetStreamParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStreamResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  sourceUrl: zod.string(),
  primaryStreamKey: zod.string(),
  backupStreamKey: zod.string().nullish(),
  rtmpsUrl: zod.string(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  switchInterval: zod.number(),
  copyrightProtection: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update stream
 */
export const UpdateStreamParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateStreamBody = zod.object({
  name: zod.string().optional(),
  sourceUrl: zod.string().optional(),
  primaryStreamKey: zod.string().optional(),
  backupStreamKey: zod.string().optional(),
  rtmpsUrl: zod.string().optional(),
  switchInterval: zod.number().optional(),
  copyrightProtection: zod.boolean().optional(),
});

export const UpdateStreamResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  sourceUrl: zod.string(),
  primaryStreamKey: zod.string(),
  backupStreamKey: zod.string().nullish(),
  rtmpsUrl: zod.string(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  switchInterval: zod.number(),
  copyrightProtection: zod.boolean(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete stream
 */
export const DeleteStreamParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteStreamResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Start streaming
 */
export const StartStreamParams = zod.object({
  id: zod.coerce.number(),
});

export const StartStreamResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  message: zod.string(),
  pid: zod.number().nullish(),
});

/**
 * @summary Stop streaming
 */
export const StopStreamParams = zod.object({
  id: zod.coerce.number(),
});

export const StopStreamResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  message: zod.string(),
  pid: zod.number().nullish(),
});

/**
 * @summary Get stream status
 */
export const GetStreamStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStreamStatusResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  message: zod.string(),
  pid: zod.number().nullish(),
});

/**
 * @summary Switch between primary and backup stream keys
 */
export const SwitchStreamKeyParams = zod.object({
  id: zod.coerce.number(),
});

export const SwitchStreamKeyResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["idle", "streaming", "error"]),
  activeKey: zod.enum(["primary", "backup"]),
  message: zod.string(),
  pid: zod.number().nullish(),
});