# lib/api-zod/src/generated/types/accessCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCode {
  id: number;
  code: string;
  used: boolean;
  usedAt?: string | null;
  createdAt: string;
}