# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginResponseRole } from "./loginResponseRole";

export interface LoginResponse {
  token: string;
  role: LoginResponseRole;
  message: string;
}