# lib/api-zod/src/generated/types/stream.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */
import type { StreamActiveKey } from "./streamActiveKey";
import type { StreamStatus } from "./streamStatus";

export interface Stream {
  id: number;
  name: string;
  sourceUrl: string;
  primaryStreamKey: string;
  backupStreamKey?: string | null;
  rtmpsUrl: string;
  status: StreamStatus;
  activeKey: StreamActiveKey;
  switchInterval: number;
  copyrightProtection: boolean;
  createdAt: string;
  updatedAt: string;
}