# lib/api-zod/src/generated/types/streamStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */
import type { StreamStatusResponseActiveKey } from "./streamStatusResponseActiveKey";
import type { StreamStatusResponseStatus } from "./streamStatusResponseStatus";

export interface StreamStatusResponse {
  id: number;
  status: StreamStatusResponseStatus;
  activeKey: StreamStatusResponseActiveKey;
  message: string;
  pid?: number | null;
}