# lib/api-zod/src/generated/types/updateStreamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Facebook Live Streamer API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStreamRequest {
  name?: string;
  sourceUrl?: string;
  primaryStreamKey?: string;
  backupStreamKey?: string;
  rtmpsUrl?: string;
  switchInterval?: number;
  copyrightProtection?: boolean;
}